requirements.txt
APScheduler==3.9.1
async-generator==1.10
attrs==21.4.0
black==22.3.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
coverage==6.3.2
cryptography==36.0.2
cycler==0.11.0
dnspython==2.2.1
email-validator==1.1.3
Flask==2.1.1
Flask-APScheduler==1.12.3
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
fonttools==4.32.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mypy-extensions==0.4.3
numpy==1.22.3
outcome==1.1.0
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.35
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
webdriver-manager==3.5.4
Werkzeug==2.1.1
wsproto==1.1.0
WTForms==3.0.1
zipp==3.8.0
